Remove hiredis support
