+++ README.md
@@ -4,7 +4,7 @@
 
 # redis-parser
 
-A high performance redis parser solution built for [node_redis](https://github.com/NodeRedis/node_redis) and [ioredis](https://github.com/luin/ioredis).
+A high performance javascript redis parser built for [node_redis](https://github.com/NodeRedis/node_redis) and [ioredis](https://github.com/luin/ioredis).
 
 Generally all [RESP](http://redis.io/topics/protocol) data will be properly parsed by the parser.
 
@@ -30,8 +30,7 @@ new Parser(options);
 * `returnError`: *function*; mandatory
 * `returnFatalError`: *function*; optional, defaults to the returnError function
 * `returnBuffers`: *boolean*; optional, defaults to false
-* `name`: *'javascript'|'hiredis'|'auto'|null*; optional, defaults to hiredis and falls back to the js parser if not available or if the stringNumbers option is choosen. Setting this to 'auto' or null is going to automatically determine what parser is available and chooses that one.
-* `stringNumbers`: *boolean*; optional, defaults to false. This is only available for the javascript parser at the moment!
+* `stringNumbers`: *boolean*; optional, defaults to false
 
 ### Example
 
@@ -55,8 +54,7 @@ var parser = new Parser({
     },
     returnFatalError: function (err) {
         lib.returnFatalError(err);
-    },
-    name: 'auto' // This returns either the hiredis or the js parser instance depending on what's available
+    }
 });
 
 Library.prototype.streamHandler = function () {
@@ -82,7 +80,6 @@ var parser = new Parser({
     returnError: function(err) {
         lib.returnError(err);
     },
-    name: 'javascript', // Use the Javascript parser
     stringNumbers: true, // Return all numbers as string instead of a js number
     returnBuffers: true // All strings are returned as buffer e.g. <Buffer 48 65 6c 6c 6f>
 });
@@ -90,22 +87,14 @@ var parser = new Parser({
 // The streamHandler as above
 ```
 
-## Further info
-
-The [hiredis](https://github.com/redis/hiredis) parser is still the fasted parser for
-Node.js and therefor used as default in redis-parser if the hiredis parser is available.
-
-Otherwise the pure js NodeRedis parser is chosen that is almost as fast as the
-hiredis parser besides some situations in which it'll be a bit slower.
-
 ## Protocol errors
 
-To handle protocol errors (this is very unlikely to happen) gracefully you should add the returnFatalError option, reject any still running command (they might have been processed properly but the reply is just wrong), destroy the socket and reconnect.
+To handle protocol errors (this is very unlikely to happen) gracefully you should add the returnFatalError option, reject any still running command (they might have been processed properly but the reply is just wrong), destroy the socket and reconnect. Be aware that while doing this, no new command may be added, so all new commands have to be buffered in the meanwhile.
 Otherwise a chunk might still contain partial data of a following command that was already processed properly but answered in the same chunk as the command that resulted in the protocol error.
 
 ## Contribute
 
-The js parser is already optimized but there are likely further optimizations possible.
+The parser is already optimized but there are likely further optimizations possible.
 
 ```
 npm install
